+++ src/domain/news/feed/command.py
@@ -11,9 +11,8 @@ from sqlmodel import select
 
 from infra.storage.rds import initialize_database, session_scope
 
-from .convert import feed_to_record, record_to_feed
 from .model import FeedItem, FeedRecord
-from .service import create_feed
+from .service import create_feed, feed_to_record, record_to_feed
 
 
 def store_feed(feed: FeedItem) -> FeedItem:
